Check plots directory exists before accessing

+++ mandelbrot.py
@@ -158,6 +158,9 @@ def mandelbrot_image_pil(data, tile, outputPath):
     #Create image and flip
     image = Image.fromarray(data)
     image = image.transpose(Image.FLIP_TOP_BOTTOM)
+
+    if not os.path.exists('plots'):
+        os.mkdir('plots')
     image.save(outputPath)
     print(f'Rendered region {tile.id}/{tile.totalTiles}\n')
 
@@ -228,10 +231,11 @@ def main():
     }
 
     #Cleanup old plots
-    oldPlots = os.listdir('plots')
-    if oldPlots != []:
-        for file in oldPlots:
-          os.remove(f'plots/{file}')
+    if os.path.isdir('plots'):
+        oldPlots = os.listdir('plots')
+        if oldPlots != []:
+            for file in oldPlots:
+                os.remove(f'plots/{file}')
 
     mandelbrot_prepare(coords['default'])
